moved common parts of tests to conftest.py, made session scope fixtures
