+++ tests/conftest.py
@@ -1,6 +1,13 @@
 # Based on https://docs.pytest.org/en/latest/example/simple.html#control-skipping-of-tests-according-to-command-line-option  # noqa: E501
 
+from mock import Mock
 import pytest
+
+from blueapi.core.context import BlueskyContext
+from blueapi.worker.reworker import RunEngineWorker
+from blueapi.service.handler import Handler, get_handler
+from blueapi.service.main import app
+from fastapi.testclient import TestClient
 
 
 def pytest_addoption(parser):
@@ -22,3 +29,36 @@ def pytest_collection_modifyitems(config, items):
         for item in items:
             if "stomp" in item.keywords:
                 item.add_marker(skip_stomp)
+
+
+class MockHandler(Handler):
+    context: BlueskyContext
+    worker: RunEngineWorker
+
+    def __init__(self) -> None:
+        self.context = Mock()
+        self.worker = Mock()
+
+    def start(self):
+        return None
+
+
+class Client:
+    def __init__(self, handler: MockHandler) -> None:
+        """Create tester object"""
+        self.handler = handler
+
+    @property
+    def client(self) -> TestClient:
+        app.dependency_overrides[get_handler] = lambda: self.handler
+        return TestClient(app)
+
+
+@pytest.fixture(scope="session")
+def handler() -> MockHandler:
+    return MockHandler()
+
+
+@pytest.fixture(scope="session")
+def client(handler: MockHandler) -> TestClient:
+    return Client(handler).client
